Remove mbe_matrix_t type.
The struct mbe_matrix_t is defined in a layout accoriding definition
of Cairo.  It is not feasible for all graphic engine.  Now, we start
supporting of other graphic engine; Skia, so use type of (co_aix[6])
defined and used by MadButterfly intead of mbe_matrix_t.  And, bridges
of graphic engines are responsible for translate matrics into the
matrix type of the graphic enigne.