Fix a potential bug when destroy a rdman.
When a rdman is dirty, free shapes and coords works specially.
Objects are append to a free list.  They are not real freed until
rdman being clean.  redraw_man_destroy() free shapes and coords with
free functions of them.  If rdman is dirty when destroy it, objects
would be leaked.  The changeset make rdman clean before free shapes
and coords to make objects being freed correctly.