Fix issue of seg. fault at xnjsmb_coord_invalidate_subtree().
It caused by _xnjsmb_coord_clone_from_subtree_mod() does not create an
JS wrapper object every descendants of the subtree.  We do not provide
any method for traveling tree.  So, it is fine to clone a subtree
without creating wrapper objects for descendants.  So, we check if an
object own a JS wrapper object before invalidating it.